services/billing-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
reportlab==4.0.7
stripe==7.4.0
razorpay==1.3.0
requests==2.31.0
httpx==0.25.2
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
slowapi==0.1.9
